+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Export `decode_error_reason()` type.
 
 ## [1.0.0] (2021-06-14)
 No change.